requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Cors==3.0.3
flask-restful
gunicorn==19.7.1
pylint==1.8.2
pymongo==3.3.1
pytest==3.0.6
psycopg2-binary==2.8.6
